Refactored config implementations
