+++ dlasset/config/__init__.py
@@ -1,0 +1,3 @@
+"""Implementations for the downloader config."""
+from .load import load_config
+from .model import Config
