+++ dlasset/config/model/__init__.py
@@ -1,0 +1,2 @@
+"""Config data model classes."""
+from .main import Config
