+++ dlasset/config/model/paths.py
@@ -1,15 +1,11 @@
-"""Implementations to load the config file."""
-import json
-import os.path
+"""Config path model class."""
+import os
 from dataclasses import InitVar, dataclass, field
 from typing import Any, cast
 
-import yaml
-from jsonschema import ValidationError, validate
-
 from dlasset.enums import Locale
 
-__all__ = ("load_config", "Config")
+__all__ = ("Paths",)
 
 
 @dataclass
@@ -38,30 +34,3 @@ class Paths:
     def lib_decrypt_dll_path(self) -> str:
         """Path of the DLL for decryption."""
         return os.path.join(self.lib, "decrypt", "Decrypt.dll")
-
-
-@dataclass
-class Config:
-    """Asset downloader config."""
-
-    paths: Paths
-
-
-def load_config(path: str) -> Config:
-    """
-    Load and validate the config.
-
-    Raises :class:`ValueError` if the config schema doesn't match.
-    """
-    with open(path, encoding="utf-8") as f:
-        config = cast(dict[Any, Any], yaml.safe_load(f))
-
-    with open("config.schema.json", encoding="utf-8") as f:
-        schema = cast(dict[Any, Any], json.load(f))
-
-    try:
-        validate(instance=config, schema=schema)
-    except ValidationError as ex:
-        raise ValueError("Config validation failed") from ex
-
-    return Config(paths=Paths(config["paths"]))
